Fixed linting errors, code style.
Not sure if you will accept this pull request, but I thought it would be better to improve the syntax style (e.g. opening curly braces are generally better not on a new line in JS because of potential errors). Also, there were some unnecessary semi-colons and at least one missing colon, and a double equals that should be triple equals.